+++ Tests/MongoSwiftTests/MongoClientTests.swift
@@ -55,36 +55,36 @@ final class MongoClientTests: MongoSwiftTestCase {
         expect(try Version("3.6.1.1")).to(equal(three61))
 
         // lt
-        expect(three6.isLessThan(three6)).to(beFalse())
-        expect(three6.isLessThan(three61)).to(beTrue())
-        expect(three61.isLessThan(three6)).to(beFalse())
-        expect(three61.isLessThan(three7)).to(beTrue())
-        expect(three7.isLessThan(three6)).to(beFalse())
-        expect(three7.isLessThan(three61)).to(beFalse())
+        expect(three6 < three6).to(beFalse())
+        expect(three6 < three61).to(beTrue())
+        expect(three61 < three6).to(beFalse())
+        expect(three61 < three7).to(beTrue())
+        expect(three7 < three6).to(beFalse())
+        expect(three7 < three61).to(beFalse())
 
         // lte
-        expect(three6.isLessThanOrEqualTo(three6)).to(beTrue())
-        expect(three6.isLessThanOrEqualTo(three61)).to(beTrue())
-        expect(three61.isLessThanOrEqualTo(three6)).to(beFalse())
-        expect(three61.isLessThanOrEqualTo(three7)).to(beTrue())
-        expect(three7.isLessThanOrEqualTo(three6)).to(beFalse())
-        expect(three7.isLessThanOrEqualTo(three61)).to(beFalse())
+        expect(three6 <= three6).to(beTrue())
+        expect(three6 <= three61).to(beTrue())
+        expect(three61 <= three6).to(beFalse())
+        expect(three61 <= three7).to(beTrue())
+        expect(three7 <= three6).to(beFalse())
+        expect(three7 <= three61).to(beFalse())
 
         // gt
-        expect(three6.isGreaterThan(three6)).to(beFalse())
-        expect(three6.isGreaterThan(three61)).to(beFalse())
-        expect(three61.isGreaterThan(three6)).to(beTrue())
-        expect(three61.isGreaterThan(three7)).to(beFalse())
-        expect(three7.isGreaterThan(three6)).to(beTrue())
-        expect(three7.isGreaterThan(three61)).to(beTrue())
+        expect(three6 > three6).to(beFalse())
+        expect(three6 > three61).to(beFalse())
+        expect(three61 > three6).to(beTrue())
+        expect(three61 > three7).to(beFalse())
+        expect(three7 > three6).to(beTrue())
+        expect(three7 > three61).to(beTrue())
 
         // gte
-        expect(three6.isGreaterThanOrEqualTo(three6)).to(beTrue())
-        expect(three6.isGreaterThanOrEqualTo(three61)).to(beFalse())
-        expect(three61.isGreaterThanOrEqualTo(three6)).to(beTrue())
-        expect(three61.isGreaterThanOrEqualTo(three7)).to(beFalse())
-        expect(three7.isGreaterThanOrEqualTo(three6)).to(beTrue())
-        expect(three7.isGreaterThanOrEqualTo(three61)).to(beTrue())
+        expect(three6 >= three6).to(beTrue())
+        expect(three6 >= three61).to(beFalse())
+        expect(three61 >= three6).to(beTrue())
+        expect(three61 >= three7).to(beFalse())
+        expect(three7 >= three6).to(beTrue())
+        expect(three7 >= three61).to(beTrue())
 
         // invalid strings
         expect(try Version("hi")).to(throwError())
